Fix: Complete workflow implementation - form generation, configuration loading, and save functionality
- Fixed nested object handling in FormGenerator for proper form generation
- Implemented configuration loading with form population
- Implemented save functionality with structure preservation
- Fixed file dialog hanging issue on Linux by using Qt dialog instead of native
- Added comprehensive state tracking for loaded schema and configuration files
- Enhanced error handling and user feedback throughout
- Added integration test suite for main workflow

Issues Fixed:
  - Form generation now shows input fields for nested objects
  - Configuration loading now populates form fields correctly
  - Save button now writes configuration to file with proper structure
  - File dialogs are now reliable and responsive on Linux

Files Modified:
  - src/ui/form_generator.h/cpp: Added nested object support
  - src/ui/main_window.h/cpp: Implemented complete workflow
  - src/io/json_reader.h/cpp: Enhanced error handling
  - src/io/json_writer.h/cpp: Enhanced error handling
  - src/io/yaml_reader.h/cpp: Enhanced error handling
  - src/io/yaml_writer.h/cpp: Enhanced error handling
  - src/core/configuration_data.cpp: Minor fixes
  - src/core/schema_validator.cpp: Minor fixes
  - src/validators/pattern_validator.h/cpp: Minor fixes
  - src/utils/string_utils.cpp: Minor fixes
  - tests/CMakeLists.txt: Added new test target
  - tests/integration/test_main_workflow.cpp: NEW comprehensive test suite

Documentation Added:
  - QUICK_TEST_GUIDE.md: Step-by-step testing instructions
  - ISSUES_FIXED_SUMMARY.md: Summary of all issues and fixes
  - FILE_DIALOG_HANG_FIX.md: Explanation of file dialog fix
  - CODE_CHANGES_DETAILED.md: Before/after code comparison
  - SESSION_COMPLETION_REPORT.md: Complete implementation report
  - QUICK_REFERENCE.md: Quick reference card
  - BUILD_AND_TEST_GUIDE.md: Build and testing guide

Resources Added:
  - resources/schemas/application-schema.json: Test schema
  - resources/configs/sample-config.json: Test configuration
  - run-configgui.sh: Helper script to run application

+++ src/utils/string_utils.cpp
@@ -38,7 +38,7 @@ std::vector<std::string> StringUtils::split(const std::string& str, char delimit
 
     // OPTIMIZATION: Pre-allocate based on delimiter count estimate
     // Count delimiters to estimate number of parts (typical +1 for last part)
-    size_t delimiter_count = std::count(str.begin(), str.end(), delimiter);
+    size_t delimiter_count = static_cast<size_t>(std::count(str.begin(), str.end(), delimiter));
     result.reserve(delimiter_count + 1);
 
     // Pre-allocate current string to avoid repeated allocations during concatenation
